build: Refactor CI scripts
Move common CI functions to common_ci.py.
